Position ok

+++ app/src/main/java/fr/univorleans/etu/renardez/kodapp/entities/daos/PositionUserDao.java
@@ -20,7 +20,7 @@ public interface PositionUserDao {
     List<PositionUser> getSomePu(int... ids);
 
     @Insert
-    int insertPos(PositionUser positionUser);
+    long insertPos(PositionUser positionUser);
 
     @Update
     int updatePos(PositionUser positionUser);
